♻️ refactor: 투표 수 표시 수정
